+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-23T08:03:01.832Z")
+    created: Optional[datetime] = parse_datetime("2025-10-23T11:44:41.290Z")
     r"""The date and time the resource was created"""